+++ packages/fs-OPFS/tsconfig.json
@@ -5,6 +5,6 @@
     "lib": ["ESNext", "DOM", "DOM.Iterable"],
     "baseUrl": "."
   },
-  "include": ["src/index.ts"],
+  "include": ["src/**/*"],
   "exclude": ["node_modules", "dist"]
 }
